break out shared extra-minter-details helper functions
